backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.46
flask-restful==0.3.9
flask-cors==3.0.10
marshmallow==3.14.1
python-dotenv==0.19.0
sympy==1.13.1
pymysql==1.0.2
cryptography==41.0.5
PyJWT==2.8.0
requests==2.28.1